+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added requirement doc for ThermoFisher Genesys 30 adapter
 - Added support for xlsx exports in Unchained Lunatic parser
 - Add column special character normalization to pandas util (and so most parsers)
+- Added support for multiple read modes in Agilent Gen5 Adapter
 
 ### Fixed